requirements.txt
fastapi==0.104.1
fastapi_utils==0.8.0
uvicorn==0.24.0
motor==3.6.0
PyJWT==2.9.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
typing_inspect==0.9.0
httpx==0.25.2
bcrypt==4.1.2